also support copying tileset fields
